requirements.txt
streamlit
python-dotenv
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
pypdf





